Delete git submodules.
There is no good reason to register the dependencies as git submodules.

It's a huge maintenance burden to sync the git submodules with those of Nyxt.

The possibility to run a CI routine within this repository is lost.  A simple
way to recover it is by adding the library to Quicklisp and running the tests
via Roswell.  Note that, as of today, the main goal is to test the library
within Nyxt's development environment.